third_party/blink/web_tests/platform/generic/media/video-buffered-too-few-arguments-expected.txt

Test how buffered.start()/end() react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS v.buffered.start() threw exception TypeError: Failed to execute 'start' on 'TimeRanges': 1 argument required, but only 0 present..
PASS v.buffered.end() threw exception TypeError: Failed to execute 'end' on 'TimeRanges': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
